Clean up main.js
